+++ packages/fast-crud/src/components/crud/fs-form-item.vue
@@ -23,7 +23,7 @@
       >
         <template #[ui.tooltip.content]>
           <span class="fs-form-helper-tooltip">
-            <fs-form-helper :helper="item.helper" :scope="scopeComputed" />
+            <fs-form-helper :helper="item.helper" :scope="scopeFunc()" />
           </span>
         </template>
         <template #[ui.tooltip.trigger]>
@@ -34,42 +34,41 @@
       </component>
     </template>
     <div class="fs-form-item-content">
-      <fs-render v-if="item.topRender" :render-func="item.topRender" :scope="scopeComputed" />
+      <fs-render v-if="item.topRender" :render-func="item.topRender" :scope="scopeFunc()" />
       <div class="fs-form-item-render">
-        <fs-render v-if="item.prefixRender" :render-func="item.prefixRender" :scope="scopeComputed" />
+        <fs-render v-if="item.prefixRender" :render-func="item.prefixRender" :scope="scopeFunc()" />
         <div class="fs-form-item-component">
-          <fs-slot-render v-if="formSlot" :slots="formSlot" :scope="scopeComputed" />
+          <fs-slot-render v-if="formSlot" :slots="formSlot" :scope="scopeFunc()" />
           <template v-else-if="item.component?.show !== false">
             <fs-render
               v-if="item.render || item.component?.render"
               :render-func="item.render || item.component.render"
-              :scope="scopeComputed"
+              :scope="scopeFunc()"
             />
             <fs-component-render
               v-else
               ref="componentRenderRef"
               v-bind="item.component"
               :model-value="modelValue"
-              :scope="scopeComputed"
+              :scope="scopeFunc()"
               @update:modelValue="updateModelValue"
             />
           </template>
         </div>
-        <fs-render v-if="item.suffixRender" :render-func="item.suffixRender" :scope="scopeComputed" />
+        <fs-render v-if="item.suffixRender" :render-func="item.suffixRender" :scope="scopeFunc()" />
       </div>
-      <fs-render v-if="item.bottomRender" :render-func="item.bottomRender" :scope="scopeComputed" />
+      <fs-render v-if="item.bottomRender" :render-func="item.bottomRender" :scope="scopeFunc()" />
       <template v-if="item.helper && computedHelperPosition !== 'label'">
-        <fs-form-helper :helper="item.helper" :scope="scopeComputed" />
+        <fs-form-helper :helper="item.helper" :scope="scopeFunc()" />
       </template>
     </div>
   </component>
 </template>
 <script lang="ts">
 import { ref, computed, defineComponent, Ref, PropType } from "vue";
-import _ from "lodash-es";
 import FsRender from "../render/fs-render.js";
 import { ScopeContext } from "../../d";
-import { useUi } from "../../use";
+import { useMerge, useUi } from "../../use";
 /**
  * form-item组件封装
  */
@@ -106,11 +105,14 @@ export default defineComponent({
   emits: ["update:modelValue"],
   setup(props, ctx) {
     const { ui } = useUi();
+    const { merge } = useMerge();
     const componentRenderRef = ref();
-    const scopeComputed: Ref<ScopeContext> = computed(() => {
+    // const scopeComputed: Ref<ScopeContext> = computed(() => {});
+
+    const scopeFunc = () => {
       const scope = props.getContextFn ? props.getContextFn() : {};
       return { value: props.modelValue, key: props.item.key, ...scope };
-    });
+    };
 
     function updateModelValue(value: any) {
       ctx.emit("update:modelValue", value);
@@ -126,7 +128,7 @@ export default defineComponent({
       return props.item?.helper?.position || (props.helper as any)?.position;
     });
     const computedHelperTooltip = computed(() => {
-      return _.merge({}, props.item.helper?.tooltip, (props.helper as any)?.tooltip);
+      return merge({}, props.item.helper?.tooltip, (props.helper as any)?.tooltip);
     });
     const computedKey = computed(() => {
       if (props.item == null) {
@@ -146,13 +148,13 @@ export default defineComponent({
     });
 
     const computedLabelRender = () => {
-      return computedLabel.value(scopeComputed.value);
+      return computedLabel.value(scopeFunc());
     };
 
     return {
       ui,
       updateModelValue,
-      scopeComputed,
+      scopeFunc,
       getComponentRef,
       componentRenderRef,
       computedHelperPosition,