+++ packages/fast-crud/src/use/use-columns.tsx
@@ -258,6 +258,7 @@ function buildSearchForm(baseOptions: CrudOptions, formType = "search", columnsM
 
 function buildFormOptions(crudOptions: DynamicallyCrudOptions, context?: UseFsContext): FormProps {
   const { t } = useI18n();
+  const { merge } = useMerge();
   context = context || {};
   const userOptions = merge(
     defaultCrudOptions.defaultOptions({ t }),
@@ -279,7 +280,7 @@ function buildColumns(userOptions: CrudOptions) {
 
   userOptions.table.columns = buildTableColumns(cloneDeep(columns));
   userOptions.table.columnsMap = buildTableColumnsMap({}, userOptions.table.columns);
-  _.merge(userOptions.toolbar, {
+  merge(userOptions.toolbar, {
     columnsFilter: {
       originalColumns: cloneDeep(userOptions.table.columns)
     }