+++ packages/fast-crud/src/use/use-dict-define.ts
@@ -4,7 +4,6 @@ import logger from "../utils/util.log";
 import { reactive, shallowReactive, UnwrapRef } from "vue";
 import LRU from "lru-cache";
 import { UnwrapNestedRefs } from "vue";
-
 const DictGlobalCache = new LRU<string, any>({
   max: 500,
   maxSize: 5000,
@@ -318,7 +317,8 @@ export class Dict<T = any> extends UnMergeable implements DictOptions<T> {
       try {
         cached.loaded = false;
         cached.loading = true;
-        const dictData = await getFromRemote();
+        let dictData = await getFromRemote();
+        dictData = dictData || [];
         if (!(dictData instanceof Array)) {
           logger.warn("dict data 格式有误，期望格式为数组，实际格式为：", dictData);
         }