+++ packages/fast-extends/src/index.umd.ts
@@ -1,5 +1,0 @@
-export * from "./uploader/index.umd";
-export * from "./editor/index.umd";
-export * from "./json/index.umd";
-export * from "./copyable/index.umd";
-export * from "./time/index.umd";