+++ packages/ui/ui-element/vite.config.js
@@ -17,7 +17,7 @@ export default ({ command, mode }) => {
     build: {
       lib: {
         entry: resolve(__dirname, "src/index.ts"),
-        name: "index"
+        name: "FsUiElement"
       },
       // cssCodeSplit: true,
       sourcemap: "inline",
@@ -56,9 +56,9 @@ export default ({ command, mode }) => {
             vue: "Vue",
             "@fast-crud/fast-crud": "FastCrud",
             "element-plus": "ElementPlus",
-            "@element-plus/icons-vue": "IconsVue",
+            "@element-plus/icons-vue": "ElementPlusIconsVue",
             "lodash-es": "_",
-            "@fast-crud/ui-interface": "UiInterface"
+            "@fast-crud/ui-interface": "FsUiInterface"
           }
         }
       }