+++ packages/ui/ui-interface/vite.config.js
@@ -18,7 +18,7 @@ export default ({ command, mode }) => {
     build: {
       lib: {
         entry: resolve(__dirname, "src/index.ts"),
-        name: "antdv"
+        name: "FsUiInterface"
       },
       // cssCodeSplit: true,
       sourcemap: "inline",
@@ -42,15 +42,12 @@ export default ({ command, mode }) => {
         ],
         // make sure to externalize deps that shouldn't be bundled
         // into your library
-        external: ["vue", "@fast-crud/fast-crud", "ant-design-vue", "@ant-design/icons-vue", "lodash-es"],
+        external: ["vue", "lodash-es"],
         output: {
           // Provide global variables to use in the UMD build
           // for externalized deps
           globals: {
             vue: "Vue",
-            "@fast-crud/fast-crud": "FastCrud",
-            "ant-design-vue": "Antdv",
-            "@ant-design/icons-vue": "IconsVue",
             "lodash-es": "_"
           }
         }