+++ packages/ui/ui-naive/vite.config.js
@@ -17,7 +17,7 @@ export default ({ command, mode }) => {
     build: {
       lib: {
         entry: resolve(__dirname, "src/index.ts"),
-        name: "naive"
+        name: "FsUiNaive"
       },
       // cssCodeSplit: true,
       sourcemap: "inline",
@@ -55,11 +55,10 @@ export default ({ command, mode }) => {
           globals: {
             vue: "Vue",
             "@fast-crud/fast-crud": "FastCrud",
-            "ant-design-vue": "Antdv",
-            "naive-ui": "Naive",
-            "@ant-design/icons-vue": "IconsVue",
+            "naive-ui": "naive",
+            "@ant-design/icons-vue": "AntDesignIconsVue",
             "lodash-es": "_",
-            "@fast-crud/ui-interface": "UiInterface"
+            "@fast-crud/ui-interface": "FsUiInterface"
           }
         }
       }